Split App.js into many .js files -- Wrote a Design Rationale in App.js
(Look to commments in App.js for further details)
