+++ AppScripts/Helpers.js
@@ -1,0 +1,30 @@
+function parseDateText(date) {
+
+    var months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
+    var hours = date.getHours();
+    var minutes = date.getMinutes();
+    var ampm = hours >= 12 ? 'PM' : 'AM';
+    hours = hours % 12;
+    hours = hours ? hours : 12;
+    minutes = minutes < 10 ? '0'+minutes : minutes;
+    const dateText = months[date.getMonth()] +
+                          " " + date.getDate() +
+                          ", " + date.getFullYear() +
+                          " " + hours + ":" + minutes +
+                          " " + ampm;
+    return dateText;
+}
+  
+  function parseSimpleDateText(date) {
+  
+    var months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
+    var hours = date.getHours();
+    var minutes = date.getMinutes();
+    var ampm = hours >= 12 ? 'PM' : 'AM';
+    hours = hours % 12;
+    hours = hours ? hours : 12;
+    minutes = minutes < 10 ? '0'+minutes : minutes;
+    const dateText = months[date.getMonth()] +
+                          " " + date.getDate();
+    return dateText;
+}
